requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appnope==0.1.3
argilla==1.4.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
black==23.1.0
blobfile==2.0.1
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
Cython==0.29.33
dataclasses-json==0.5.7
debugpy==1.6.5
decorator==5.1.1
Deprecated==1.2.13
# Editable Git install with no remote (detectron2==0.6)
-e c:\users\jayson\unstructured-test\detectron2
dnspython==2.3.0
docopt==0.6.2
effdet==0.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.94.0
filelock==3.9.0
Flask==2.2.3
Flask-Script==2.0.6
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
fvcore==0.1.5.post20221221
google-auth==2.16.2
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.3
gunicorn==20.1.0
h11==0.14.0
honcho==1.1.0
html==1.13
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.11.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
iopath==0.1.9
ipykernel==6.20.1
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsons==1.6.3
jupyter_client==7.4.8
jupyter_core==5.1.3
kiwisolver==1.4.4
langchain==0.0.107
layoutparser==0.3.4
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.26.1
opencv-python==4.6.0.66
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.0
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.2.1
pipreqs==0.4.11
platformdirs==2.6.2
plotly==5.12.0
portalocker==2.7.0
prompt-toolkit==3.0.36
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycodestyle @ file:///C:/b/abs_d77nxvklcq/croot/pycodestyle_1674267231034/work
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.6
Pygments==2.14.0
pyparsing==3.0.9
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2022.7.1
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
stack-data==0.6.2
starlette==0.26.0.post1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.1.0
tiktoken==0.1.2
timm==0.6.12
tokenizers==0.13.2
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.8.1
transformers==4.25.1
typing-inspect==0.8.0
typing_extensions==4.5.0
typish==1.9.3
unstructured==0.5.3
unstructured-inference==0.2.7
urllib3==1.26.13
uvicorn==0.21.0
waitress==2.1.2
Wand==0.6.11
wcwidth==0.2.5
Werkzeug==2.2.3
win32-setctime==1.1.0
wincertstore==0.2
wrapt==1.14.1
XlsxWriter==3.0.9
yacs==0.1.8
yarg==0.1.9
yarl==1.8.2
